disable seeker for compressReader
